refactor: Extract DirtyTracking module from Canvas
Move dirty?, dirty_region, clean!, mark_dirty, expand_dirty, and
mark_clipped_dirty into ChromaWave::DirtyTracking mixin. Canvas
includes the module and calls copy_dirty_state in initialize_copy.

Reduces Canvas ClassLength from 220 to 160, allowing rubocop
ClassLength Max to drop from 220 to 190.
